extracted select adjacent handler
